Move deprecations handling to a kernel::__destruct function
