+++ src/Glpi/Http/LegacyPostRequestActionsListener.php
@@ -48,86 +48,6 @@ class LegacyPostRequestActionsListener implements EventSubscriberInterface
 
     public function onFinishRequest(): void
     {
-        $this->resetSessionAjaxParam();
-        $this->triggerGlobalsDeprecation();
-    }
-
-    private function resetSessionAjaxParam(): void
-    {
         \Session::resetAjaxParam();
-    }
-
-    private function triggerGlobalsDeprecation(): void
-    {
-        if (in_array(realpath(GLPI_ROOT . '/inc/includes.php'), get_included_files(), true)) {
-            // The following deprecations/warnings are already triggered in `inc/includes.php`.
-            return;
-        }
-
-        /**
-         * @var mixed|null $AJAX_INCLUDE
-         */
-        global $AJAX_INCLUDE;
-        if (isset($AJAX_INCLUDE)) {
-            \Toolbox::deprecated('The global `$AJAX_INCLUDE` variable usage is deprecated. Use "$this->setAjax()" from your controllers instead.');
-        }
-
-        /**
-         * @var mixed|null $SECURITY_STRATEGY
-         */
-        global $SECURITY_STRATEGY;
-        if (isset($SECURITY_STRATEGY)) {
-            // TODO Add a Route attribute to define the Firewall strategy.
-            trigger_error(
-                'The global `$SECURITY_STRATEGY` variable has no effect anymore.',
-                E_USER_WARNING
-            );
-        }
-
-        /**
-         * @var mixed|null $USEDBREPLICATE
-         * @var mixed|null $DBCONNECTION_REQUIRED
-         */
-        global $USEDBREPLICATE, $DBCONNECTION_REQUIRED;
-        if (isset($USEDBREPLICATE) || isset($DBCONNECTION_REQUIRED)) {
-            trigger_error(
-                'The global `$USEDBREPLICATE` and `$DBCONNECTION_REQUIRED` variables has no effect anymore. Use "DBConnection::getReadConnection()" to get the most apporpriate connection for read only operations.',
-                E_USER_WARNING
-            );
-        }
-
-        /**
-         * @var mixed|null $PLUGINS_EXCLUDED
-         * @var mixed|null $PLUGINS_INCLUDED
-         */
-        global $PLUGINS_EXCLUDED, $PLUGINS_INCLUDED;
-        if (isset($PLUGINS_EXCLUDED) || isset($PLUGINS_INCLUDED)) {
-            trigger_error(
-                'The global `$PLUGINS_EXCLUDED` and `$PLUGINS_INCLUDED` variables has no effect anymore.',
-                E_USER_WARNING
-            );
-        }
-
-        /**
-         * @var mixed|null $skip_db_check
-         */
-        global $skip_db_check;
-        if (isset($skip_db_check)) {
-            trigger_error(
-                'The global `$skip_db_check` variable has no effect anymore.',
-                E_USER_WARNING
-            );
-        }
-
-        /**
-         * @var mixed|null $skip_db_check
-         */
-        global $dont_check_maintenance_mode;
-        if (isset($dont_check_maintenance_mode)) {
-            trigger_error(
-                'The global `$dont_check_maintenance_mode` variable has no effect anymore.',
-                E_USER_WARNING
-            );
-        }
     }
 }
