+++ src/Glpi/Kernel/Kernel.php
@@ -37,6 +37,7 @@ namespace Glpi\Kernel;
 use Glpi\Config\ConfigProviderConsoleExclusiveInterface;
 use Glpi\Config\ConfigProviderWithRequestInterface;
 use Glpi\Config\LegacyConfigProviders;
+use Symfony\Component\HttpFoundation\RequestStack;
 use Symfony\Component\HttpKernel\Kernel as BaseKernel;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 use Symfony\Component\Routing\Loader\Configurator\RoutingConfigurator;
@@ -51,6 +52,11 @@ final class Kernel extends BaseKernel
     {
         $env ??= $_ENV['GLPI_ENVIRONMENT_TYPE'] ?? $_SERVER['GLPI_ENVIRONMENT_TYPE'] ?? 'production';
         parent::__construct($env, $env === 'production');
+    }
+
+    public function __destruct()
+    {
+        $this->triggerGlobalsDeprecation();
     }
 
     public function getProjectDir(): string
@@ -115,5 +121,79 @@ final class Kernel extends BaseKernel
     protected function configureRoutes(RoutingConfigurator $routes): void
     {
         $routes->import($this->getProjectDir() . '/src/Glpi/Controller', 'attribute');
+    }
+
+    private function triggerGlobalsDeprecation(): void
+    {
+        if (in_array(GLPI_ROOT . '/inc/includes.php', get_included_files(), true)) {
+            // The following deprecations/warnings are already triggered in `inc/includes.php`.
+            return;
+        }
+
+        /**
+         * @var mixed|null $AJAX_INCLUDE
+         */
+        global $AJAX_INCLUDE;
+        if (isset($AJAX_INCLUDE)) {
+            \Toolbox::deprecated('The global `$AJAX_INCLUDE` variable usage is deprecated. Use "$this->setAjax()" from your controllers instead.');
+        }
+
+        /**
+         * @var mixed|null $SECURITY_STRATEGY
+         */
+        global $SECURITY_STRATEGY;
+        if (isset($SECURITY_STRATEGY)) {
+            // TODO Add a Route attribute to define the Firewall strategy.
+            trigger_error(
+                'The global `$SECURITY_STRATEGY` variable has no effect anymore.',
+                E_USER_WARNING
+            );
+        }
+
+        /**
+         * @var mixed|null $USEDBREPLICATE
+         * @var mixed|null $DBCONNECTION_REQUIRED
+         */
+        global $USEDBREPLICATE, $DBCONNECTION_REQUIRED;
+        if (isset($USEDBREPLICATE) || isset($DBCONNECTION_REQUIRED)) {
+            trigger_error(
+                'The global `$USEDBREPLICATE` and `$DBCONNECTION_REQUIRED` variables has no effect anymore. Use "DBConnection::getReadConnection()" to get the most apporpriate connection for read only operations.',
+                E_USER_WARNING
+            );
+        }
+
+        /**
+         * @var mixed|null $PLUGINS_EXCLUDED
+         * @var mixed|null $PLUGINS_INCLUDED
+         */
+        global $PLUGINS_EXCLUDED, $PLUGINS_INCLUDED;
+        if (isset($PLUGINS_EXCLUDED) || isset($PLUGINS_INCLUDED)) {
+            trigger_error(
+                'The global `$PLUGINS_EXCLUDED` and `$PLUGINS_INCLUDED` variables has no effect anymore.',
+                E_USER_WARNING
+            );
+        }
+
+        /**
+         * @var mixed|null $skip_db_check
+         */
+        global $skip_db_check;
+        if (isset($skip_db_check)) {
+            trigger_error(
+                'The global `$skip_db_check` variable has no effect anymore.',
+                E_USER_WARNING
+            );
+        }
+
+        /**
+         * @var mixed|null $skip_db_check
+         */
+        global $dont_check_maintenance_mode;
+        if (isset($dont_check_maintenance_mode)) {
+            trigger_error(
+                'The global `$dont_check_maintenance_mode` variable has no effect anymore.',
+                E_USER_WARNING
+            );
+        }
     }
 }
